+++ common/src/main/java/com/github/teamfossilsarcheology/fossil/block/custom_blocks/FossilBlock.java
@@ -3,13 +3,19 @@ package com.github.teamfossilsarcheology.fossil.block.custom_blocks;
 import com.github.teamfossilsarcheology.fossil.entity.prehistoric.base.PrehistoricEntityInfo;
 import com.github.teamfossilsarcheology.fossil.tags.ModItemTags;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.storage.loot.LootContext;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 public class FossilBlock extends Block {
 
@@ -19,8 +25,9 @@ public class FossilBlock extends Block {
 
     @Override
     public @NotNull List<ItemStack> getDrops(BlockState state, LootContext.Builder builder) {
-        int i = builder.getLevel().random.nextInt(PrehistoricEntityInfo.entitiesWithBones().size());
-        PrehistoricEntityInfo info = PrehistoricEntityInfo.entitiesWithBones().get(i);
+        List<ItemStack> items = new ArrayList<>();
+        int idx = builder.getLevel().random.nextInt(PrehistoricEntityInfo.entitiesWithBones().size());
+        PrehistoricEntityInfo info = PrehistoricEntityInfo.entitiesWithBones().get(idx);
         builder = builder
                 .withDynamicDrop(ModItemTags.ARM_BONES.location(), (a, c) -> c.accept(new ItemStack(info.armBoneItem)))
                 .withDynamicDrop(ModItemTags.FOOT_BONES.location(), (a, c) -> c.accept(new ItemStack(info.footBoneItem)))
@@ -30,7 +37,16 @@ public class FossilBlock extends Block {
                 .withDynamicDrop(ModItemTags.TAIL_BONES.location(), (a, c) -> c.accept(new ItemStack(info.tailBoneItem)))
                 .withDynamicDrop(ModItemTags.UNIQUE_BONES.location(), (a, c) -> c.accept(new ItemStack(info.uniqueBoneItem)))
                 .withDynamicDrop(ModItemTags.VERTEBRAE_BONES.location(), (a, c) -> c.accept(new ItemStack(info.vertebraeBoneItem)));
-        return super.getDrops(state, builder);
+        Map<Enchantment, Integer> enchantments = EnchantmentHelper.getEnchantments( builder.getParameter(LootContextParams.TOOL));
+        int count = 1;
+        int fortune = enchantments.getOrDefault(Enchantments.BLOCK_FORTUNE, 0);
+        if (fortune > 0) {
+            count += builder.getLevel().random.nextInt(fortune);
+        }
+        for (int i = 0; i < count; i++) {
+            items.addAll(super.getDrops(state, builder));
+        }
+        return items;
     }
 
     @Override